Allow a custom form name
It'd be nice to be able to pass a custom form name when saving submissions. That way the saved entries are broken up by form name in the admin.

The only way I could think of doing it was passing it through the message field though. Seems to work just fine but doesn't feel ideal.